move valkyrie replacements for hyrax/collection_behavior  to file mirroring the hyrax directory structure
* also moves convert_to_active_fedora_ids to IdConverterService.
